fix: renames codejam.js -> sumOfother.js. Fix test style.
